extract logger into own package
