+++ pkg/security/secl/rules/ruleset.go
@@ -15,6 +15,7 @@ import (
 	"github.com/spf13/cast"
 
 	"github.com/DataDog/datadog-agent/pkg/security/secl/compiler/eval"
+	"github.com/DataDog/datadog-agent/pkg/security/secl/log"
 )
 
 // MacroID represents the ID of a macro
@@ -167,7 +168,7 @@ type RuleSet struct {
 	scopedVariables  map[Scope]VariableProvider
 	// fields holds the list of event field queries (like "process.uid") used by the entire set of rules
 	fields []string
-	logger Logger
+	logger log.Logger
 	pool   *eval.ContextPool
 }
 
@@ -768,13 +769,7 @@ func (rs *RuleSet) LoadPolicies(loader *PolicyLoader, opts PolicyLoaderOpts) *mu
 
 // NewRuleSet returns a new ruleset for the specified data model
 func NewRuleSet(model eval.Model, eventCtor func() eval.Event, opts *Opts, evalOpts *eval.Opts, macroStore *eval.MacroStore) *RuleSet {
-	var logger Logger
-
-	if opts.Logger != nil {
-		logger = opts.Logger
-	} else {
-		logger = &NullLogger{}
-	}
+	logger := log.OrNullLogger(opts.Logger)
 
 	return &RuleSet{
 		model:            model,
